+++ TIDALDL-PY/tidal_dl/printf.py
@@ -28,7 +28,7 @@ from .environment import isTermux
 from .lang.language import LANG
 
 
-VERSION = '2026.9.26.0'
+VERSION = '2026.9.26.1'
 PROJECT_URL = 'https://github.com/OpenNerdz/tidekeeper'
 
 print_mutex = threading.Lock()